move dynamic triggers
