+++ functions/utils/trigger_operations.ts
@@ -4,8 +4,8 @@ import { EventTriggerResponseObject } from "deno-slack-api/typed-method-types/wo
 import PromptSurveyWorkflow from "../../workflows/prompt_survey.ts";
 import RemoveSurveyWorkflow from "../../workflows/remove_survey.ts";
 
-import promptSurveyTrigger from "../../triggers/prompt_survey_trigger.ts";
-import removeSurveyTrigger from "../../triggers/remove_survey_trigger.ts";
+import promptSurveyTrigger from "../triggers/prompt_survey_trigger.ts";
+import removeSurveyTrigger from "../triggers/remove_survey_trigger.ts";
 
 /**
  * The /utils directory exports commonly shared or abstracted
